// backend/src/api/client/ApiTypes.ts
export type CommentReactionNotification = {
  id: string;
  notificationType: 'comment-reaction-notification';
  isRead: boolean;
  actor: {
    id: string;
    firstName: string;
    lastName: string;
    avatar: string;
  };
  targetComment: {
    id: string;
    content: string;
  };
  targetPost: {
    id: string;
  };
  createdAt: string;
};
export type CommentCommentNotification = {
  id: string;
  notificationType: 'comment-comment-notification';
  isRead: boolean;
  actor: {
    id: string;
    firstName: string;
    lastName: string;
    avatar: string;
  };
  targetComment: {
    id: string;
    content: string;
  };
  targetPost: {
    id: string;
  };
  comment: {
    id: string;
    content: string;
  };
  createdAt: string;
};
export type PostCommentNotification = {
  id: string;
  notificationType: 'post-comment-notification';
  isRead: boolean;
  actor: {
    id: string;
    firstName: string;
    lastName: string;
    avatar: string;
  };
  targetPost: {
    id: string;
    postType: 'moment' | 'sprint-result';
    content?: string;
  };
  createdAt: string;
};
export type PostReactionNotification = {
  id: string;
  notificationType: 'post-reaction-notification';
  isRead: boolean;
  actor: {
    id: string;
    firstName: string;
    lastName: string;
    avatar: string;
  };
  targetPost: {
    id: string;
    postType: 'moment' | 'sprint-result';
    content?: string;
  };
  createdAt: string;
};
// ...
export type CommunityInvitationNotification = {
  id: string;
  notificationType: 'community-invitation-notification';
  isRead: boolean;
  actor: {
    id: string;
    firstName: string;
    lastName: string;
    avatar: string;
  };
  community: {
    id: string;
    name: string;
  };
  createdAt: string;
};
// ...
export type GetMyNotificationsResponse = Array<
  | CommentReactionNotification
  | CommentCommentNotification
  | PostCommentNotification
  | PostReactionNotification
  | CommunityInvitationNotification
>;